requirements.txt

yfinance>=0.2.32
pandas>=1.5.3
numpy>=1.24.0
python-dotenv>=1.0.0
requests>=2.31.0
